+++ metadata.rb
@@ -1,15 +1,15 @@
-name             'unblibraries-apache2'
-maintainer       'UNB Libraries Systems'
-maintainer_email 'jsanford@unb.ca'
+name             'lsulibraries-apache2'
+maintainer       'LSU Libraries'
+maintainer_email 'jpeak5@lsu.edu'
 license          'All rights reserved'
-description      'Apache2 webserver wrapper recipe for UNB Libraries LAMP stack'
+description      'Apache2 webserver wrapper recipe for LAMP stack'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends "unblibraries-essentials"
-depends "unblibraries-mysql"
+depends "lsulibraries-base"
+depends "lsulibraries-mysql"
 depends "php"
 depends "apache2"
-depends "unblibraries-users"
-depends "unblibraries-munin"
-depends "hostname"
+#depends "lsulibraries-users"
+#depends "lsulibraries-munin"
+#depends "hostname"